implemented new endpoints

+++ src/app/components/nav/nav.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import swal from 'sweetalert';
 
 @Component({
   selector: 'app-nav',
@@ -20,7 +21,20 @@ export class NavComponent implements OnInit {
   }
 
   logOut() {
-    this.router.navigate(['login']);
-    localStorage.clear();
+    swal({
+      title: 'Are you sure?',
+      icon: 'warning',
+      buttons: [ true, true ],
+      dangerMode: true
+    }).then((willDelete) => {
+      if (willDelete) {
+        this.router.navigate(['login']);
+        localStorage.clear();
+      } else {
+        swal('Okay!', {
+          icon: 'info'
+        });
+      }
+    });
   }
 }
